+++ src/py42/services/detectionlists/user_profile.py
@@ -1,4 +1,5 @@
 from py42.exceptions import Py42BadRequestError
+from py42.exceptions import Py42CloudAliasLimitExceededError
 from py42.exceptions import Py42NotFoundError
 from py42.services import BaseService
 
@@ -169,7 +170,12 @@ class DetectionListUserService(BaseService):
             u"cloudUsernames": [alias],
         }
         uri = self._make_uri(u"/addcloudusernames")
-        return self._connection.post(uri, json=data)
+        try:
+            return self._connection.post(uri, json=data)
+        except Py42BadRequestError as err:
+            if "Cloud usernames must be less than or equal to" in err.response.text:
+                raise Py42CloudAliasLimitExceededError(err)
+            raise err
 
     def remove_cloud_alias(self, user_id, alias):
         """Remove one or more cloud alias.