+++ tests/services/detectionlists/test_user_profile.py
@@ -3,8 +3,19 @@ from requests import Response
 from requests.exceptions import HTTPError
 
 from py42.exceptions import Py42BadRequestError
+from py42.exceptions import Py42CloudAliasLimitExceededError
 from py42.services.detectionlists.user_profile import DetectionListUserService
 from py42.services.users import UserService
+
+
+CLOUD_ALIAS_LIMIT_EXCEEDED_ERROR_MESSAGE = """{
+"pop-bulletin": {
+"type$": "com.code42.detectionlistmanagement.DetectionListMessages.ValidationError",
+"text$": "Command or Query is invalid: Cloud usernames must be less than or equal to 2",
+"details": [],
+"reason": "Cloud usernames must be less than or equal to 2"
+}
+}"""
 
 
 class TestDetectionListUserClient(object):
@@ -31,6 +42,19 @@ class TestDetectionListUserClient(object):
         user_client = UserService(mock_connection)
         response = mocker.MagicMock(spec=Response)
         response.status_code = 400
+        exception = mocker.MagicMock(spec=HTTPError)
+        exception.response = response
+        mock_connection.post.side_effect = Py42BadRequestError(exception)
+        return user_client
+
+    @pytest.fixture
+    def mock_user_client_error_on_adding_cloud_aliases(
+        self, mocker, mock_connection, user_context, py42_response
+    ):
+        user_client = UserService(mock_connection)
+        response = mocker.MagicMock(spec=Response)
+        response.status_code = 400
+        response.text = CLOUD_ALIAS_LIMIT_EXCEEDED_ERROR_MESSAGE
         exception = mocker.MagicMock(spec=HTTPError)
         exception.response = response
         mock_connection.post.side_effect = Py42BadRequestError(exception)
@@ -235,3 +259,18 @@ class TestDetectionListUserClient(object):
             posted_data["tenantId"] == user_context.get_current_tenant_id()
             and posted_data["userId"] == "942897397520289999"
         )
+
+    def test_add_cloud_alias_limit_raises_custom_error_on_limit(
+        self,
+        mock_connection,
+        user_context,
+        mock_user_client_error_on_adding_cloud_aliases,
+    ):
+        detection_list_user_client = DetectionListUserService(
+            mock_connection,
+            user_context,
+            mock_user_client_error_on_adding_cloud_aliases,
+        )
+        with pytest.raises(Py42CloudAliasLimitExceededError) as err:
+            detection_list_user_client.add_cloud_alias("942897397520289999", "Test")
+        assert "Cloud alias limit exceeded." in str(err.value)